Finished

+++ shutdown-android/app/src/main/java/org/lovekapibarasan/shutdown_android/config/RepeatingTriggerConfig.java
@@ -3,14 +3,14 @@ package org.lovekapibarasan.shutdown_android.config;
 public class RepeatingTriggerConfig {
     public Integer startHour;
     public Integer startMinute;
-    public Long intervalMinutes;
+    public Integer intervalMinutes;
     public Integer endHour;
     public Integer endMinute;
-    public Long durationMinutes;
+    public Integer durationMinutes;
 
     public RepeatingTriggerConfig(Integer startHour, Integer startMinute,
-                                  Long intervalMinutes, Integer endHour,
-                                  Integer endMinute, Long durationMinutes) {
+                                  Integer intervalMinutes, Integer endHour,
+                                  Integer endMinute, Integer durationMinutes) {
         this.startHour = startHour;
         this.startMinute = startMinute;
         this.intervalMinutes = intervalMinutes;
